HHH-8191 Teradata 14.0_A
