+++ databases/teradata13/resources/hibernate.properties
@@ -1,0 +1,23 @@
+
+hibernate.dialect org.hibernate.dialect.Teradata13Dialect
+hibernate.connection.driver_class com.teradata.jdbc.TeraDriver
+hibernate.connection.url jdbc:teradata://testDBS/XLOG=DEBUG,LOB_SUPPORT=ON,LOB_TEMP_TABLE=hibernate.JdbcLobUpdate,TMODE=ANSI
+hibernate.connection.username hibernate
+hibernate.connection.password hibernate
+
+hibernate.connection.pool_size 5
+
+hibernate.show_sql true
+hibernate.format_sql true
+
+hibernate.max_fetch_depth 5
+
+hibernate.cache.region_prefix hibernate.test
+hibernate.cache.provider_class org.hibernate.cache.HashtableCacheProvider
+hibernate.jdbc.batch_versioned_data true
+
+#needed for AstParser tests
+hibernate.connection.autocommit true
+
+
+
