+++ hibernate-core/src/main/java/org/hibernate/dialect/Dialect.java
@@ -2628,34 +2628,12 @@ public abstract class Dialect implements ConversionContext {
 		return true;
 	}
 
-
 	/**
-	 * Does this dialect require a special function for Constraint processing ?
+	 * Does locking information get appended to SQL ?
 	 *
-	 * @return boolean
-	 *
+	 * @return  boolean
 	 */
-	public boolean constraintRequiresFunction() {
-		return false;
-	}
-
-	/**
-	 * create the text of the special function needed for constraint processing
-	 *
-	 * @return String
-	 *
-	 */
-	public String[] generateConstraintFunctionSql(String alter, String tableName, String constraintName) {
-		return null;
-	}
-
-	/**
-	 * Do column names need to be specified when creating an Index ?
-	 *
-	 * @return boolean
-	 *
-	 */
-	public boolean doesIndexNameRequireColumnNames() {
-		return false;
+	public boolean isLockAppended() {
+		return true;
 	}
 }
